requirements.txt
Keras==2.2.4
Pillow==6.0.0
codecov==2.0.15
cython==0.29.10
flake8==3.7.7
future==0.17.1
hdbscan==0.8.22
hypothesis==4.24.6
imutils==0.5.2
matplotlib==3.1.0
mock==3.0.5
numba==0.44.1
numpy==1.16.4
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
pandas==0.24.2
protobuf==3.8.0
pydot==1.4.1
pytest==4.6.3
pytest-cov==2.7.1
pyyaml==4.2b2
scikit-learn==0.21.2
scipy==1.3.0
tensorflow==1.14.0
torch==1.1.0
tqdm==4.32.2
typing==3.7.4
umap-learn==0.3.9
jupyter==1.0.0
graphviz==0.11.1
Sphinx==2.1.2
docutils==0.14
recommonmark==0.5.0
